chore: refactor role definitions to separate file
